IT'S IN THE RANGE!!!!!!!!
Find range WORKS!

Insert & insert helper now adds the keys to a vector in order for find
range to work properly.

It's a perfectly working BST.

AVL tree part is not implemented. I don't want to do the turny-turns.

Will I?
I don't know. Probably not.

But I DID IT!!!!!!